------------------------------------------------------------------------ r282 | stericson.g1 | 2013-07-23 16:24:41 +0000 (Tue, 23 Jul 2013) | 2 lines
Cleanup

------------------------------------------------------------------------
r281 | stericson.g1 | 2013-07-23 16:14:36 +0000 (Tue, 23 Jul 2013) | 5 lines

Fix for a major bug that caused hanging when a large number of commands were executed.

This fix also reduces the memory footprint of the shell since it imposes the max number of commands that are allowed in the list of commands executed. Once this max is hit a cleanup is performed at which time the list of commands being executed is reduced to half of the max amount.

This is now RootTools 3.1
------------------------------------------------------------------------
r280 | stericson.g1 | 2013-07-22 14:58:18 +0000 (Mon, 22 Jul 2013) | 1 line

Added the ability to set the default timeout for ShellCommands