------------------------------------------------------------------------ r292 | thuanbkit@gmail.com | 2013-09-26 12:32:28 +0000 (Thu, 26 Sep 2013) | 1 line
Use NIO in installBinary method for faster IO
------------------------------------------------------------------------
r291 | stericson.g1 | 2013-09-25 14:32:27 +0000 (Wed, 25 Sep 2013) | 1 line

Fixed a null pointer exception that was occurring. The exact cause is still unknown. We need to implement a way to get debug information to the RootTools Development team. This is something that we can allow developers to support.